% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btn_add_custom_position.R
\name{btn_add_custom_position}
\alias{btn_add_custom_position}
\title{Add a row to the Position Table for a custom position}
\usage{
btn_add_custom_position(input, Position)
}
\arguments{
\item{input}{}
}
\description{
Allows the user to add a row within an organization
  to create a custom position. In order to add a custom position, 
  there must not be any custom position input field without a name.
}

