% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btn_configure_congregation.R
\name{btn_configure_congregation}
\alias{btn_configure_congregation}
\title{Send Configuration Details to the Database}
\usage{
btn_configure_congregation(input)
}
\arguments{
\item{input}{The input object from the Shiny session}
}
\description{
Adds or updates unit number, unit name, and unit type to the 
  Congregation table.
}

