% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btn_configureDatabase.R
\name{btn_configureDatabase}
\alias{btn_configureDatabase}
\title{Configure the Values of the Unit Database.}
\usage{
btn_configureDatabase(input)
}
\arguments{
\item{input}{The input object from the shiny session.}
}
\description{
Provide values to the unit number, unit name, unit type,
  and unit leader fields in the CONGREGATION table.
}

