% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btn_importRecords.R
\name{btn_importRecords}
\alias{btn_importRecords}
\title{Import Membership Records}
\usage{
btn_importRecords(input)
}
\arguments{
\item{input}{the input object from the shiny session}
}
\description{
Inserts or updates membership records from a CSV file into
  the Membership table.
}

