% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btn_initializeDatabase.R
\name{btn_initializeDatabase}
\alias{btn_initializeDatabase}
\title{Initialize the Values of the Unit Database.}
\usage{
btn_initializeDatabase(input)
}
\arguments{
\item{input}{The input object from the shiny session.}
}
\description{
Provide values to the unit number, unit name, unit type,
  and unit leader fields in the CONGREGATION table.
}

