% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/btn_renderReport.R
\name{btn_renderReport}
\alias{btn_renderReport}
\title{Render a Report}
\usage{
btn_renderReport(input)
}
\arguments{
\item{input}{The input object from the shiny session}
}
\description{
Renders a report
}

