% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btn_save_custom_position_changes.R
\name{btn_save_custom_position_changes}
\alias{btn_save_custom_position_changes}
\title{Save Change to Custom Positions}
\usage{
btn_save_custom_position_changes(input)
}
\arguments{
\item{input}{The input object from the shiny session}
}
\description{
Save current changes to custom positions
}

