% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initialize_database.R
\name{initialize_database}
\alias{initialize_database}
\title{Intitialize the MLS Membership Database}
\usage{
initialize_database()
}
\description{
When the application is first started, it checks to see if the
  database has previously been created.  If not, a new database is made and the
  user is prompted to load a membership file.  If the database already exists, the
  application will start as usual.
}
\author{
Benjamin Nutter
}

