% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{Launch the LDS MLS Extension Application}
\usage{
launch_application(launch_browser)
}
\description{
Start the Application.  First, the system checks that the database 
  file has been initialized in the appropriate location.  If not, it creates the
  database and tables.  After this check, the application is launched.
}
\examples{
\dontrun{
  launch_application()
}

}
\author{
Benjamin Nutter
}

