% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query_create_database.R
\name{query_create_database}
\alias{query_create_database}
\title{Create the Tables for the Application Database}
\usage{
query_create_database()
}
\description{
When \code{\link{startup_initialize_database}} runs, it executes a query
  to create the database and various tables used by the application.  The SQL
  code used to generate those tables is called by this function.
}
\author{
Benjamin Nutter
}

