% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt_baptisms.R
\name{rpt_baptisms}
\alias{rpt_baptisms}
\title{Upcoming Baptisms}
\usage{
rpt_baptisms(start_date, end_date, cat = FALSE)
}
\arguments{
\item{start_date}{The start date of the eligibility period}

\item{end_date}{The end date of the eligibility period}

\item{cat}{logical. Determines if the results should be printed with the
  \code{cat} function.}
}
\value{
Returns a table with the member's name, date of the eighth
  birthday, and sex.
}
\description{
Retrieve the membership records of individuals who will
  turn eight years old between two dates.  This provides a list of
  members who will become eligible for baptism in the specified time
  period.
}

