% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpt_youth.R
\name{rpt_youth}
\alias{rpt_youth}
\title{Upcoming Youth}
\usage{
rpt_youth(start_date, end_date, cat = FALSE)
}
\arguments{
\item{start_date}{The start date of the eligibility period}

\item{end_date}{The end date of the eligibility period}

\item{cat}{logical. Determines if the results should be printed with the
  \code{cat} function.}
}
\value{
Returns a table with the member's name, date of the eighth
  birthday, and sex.
}
\description{
Provides a table of all members turning twelve years old
  within the specified time frame.
}

