% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sqlite_exe.R
\name{sqlite_exe}
\alias{sqlite_exe}
\title{Retrieve the Location of the SQLite Executable}
\usage{
sqlite_exe()
}
\description{
Retrieve the Location of the SQLite Executable
}
\author{
Benjamin Nutter
}

