% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup_congregation_configured.R
\name{startup_congregation_configured}
\alias{startup_congregation_configured}
\title{Determine if there congregation has a unit number}
\usage{
startup_congregation_configured()
}
\description{
Returns a logical value when a single unit number
  exists in the Congregation table.
}

