% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/startup_congregationConfigured.R
\name{startup_congregationConfigured}
\alias{startup_congregationConfigured}
\title{Determine if Congregation is Configured}
\usage{
startup_congregationConfigured()
}
\description{
At a minimum, \code{ldsmls} requires that a unit number be
associated with the database.  The unit number is used in determining the
unique identifier for each member.  On start up, this function is run
to determine if 1) the Congregation table has exactly one row, or 2) the
Congregation table has one row and a value in the Unit Number field.  If
neither condition is satisfied, \code{FALSE} is returned and triggers the
UI to force the user to enter a unit number.
}

