% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_configure_position.R
\name{ui_configure_position}
\alias{ui_configure_position}
\title{User Interface for Configuring Positions}
\usage{
ui_configure_position(input, Position)
}
\arguments{
\item{input}{The input object from the shiny session}
}
\description{
Generate a table for managing positions within organizations
}

