% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_configureDatabase.R
\name{ui_configureDatabase}
\alias{ui_configureDatabase}
\title{Configure Database User Interface}
\usage{
ui_configureDatabase(input, output)
}
\arguments{
\item{input}{the input object from the shiny session}

\item{output}{the output object from the shiny session}
}
\description{
Renders the User Interface with Database Configuration
 options.  These are stored in the Congregation table.
}

