% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_congregation_parameters.R
\name{ui_congregation_parameters}
\alias{ui_congregation_parameters}
\title{Render the UI for changing congregation parameters}
\usage{
ui_congregation_parameters(Congregation)
}
\arguments{
\item{Congregation}{The Congregration data frame from the RV list}
}
\description{
Generates the UI for changing congregation parameters
  and populates it with the current values from the database.
}

