% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_importRecords.R
\name{ui_importRecords}
\alias{ui_importRecords}
\title{Import Membership Records}
\usage{
ui_importRecords(input)
}
\arguments{
\item{input}{The input object from the shiny session}
}
\description{
Imports Membership Records and assigns a unique identifier.
}

