% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_initializeDatabase.R
\name{ui_initializeDatabase}
\alias{ui_initializeDatabase}
\title{User Interface for Initializing the Database}
\usage{
ui_initializeDatabase(input, output)
}
\arguments{
\item{input}{The input object from the shiny session}

\item{output}{The output object from the shiny session}
}
\description{
This user interface is displayed on the first time the
application is run when the database needs to be configured with at
least a unit number.
}

