% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_databasePath.R
\name{util_databasePath}
\alias{util_databasePath}
\title{Retrieve path to Database File}
\usage{
util_databasePath()
}
\description{
A wrapper function for getting the path to the database
  file.
}

