% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util_sqlValuePrep.R
\name{util_sqlValuePrep}
\alias{util_sqlValuePrep}
\title{Prepare A Value for SQL Injection}
\usage{
util_sqlValuePrep(x)
}
\arguments{
\item{x}{The value to be injected into SQL.}
}
\description{
Prepares values for injection into a SQL database
  by wrapping in quotes, if necessary, or assigning missing values to
  a SQL \code{NULL}.
}

