% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btn_importRecords.R
\name{btn_importRecords}
\alias{btn_importRecords}
\title{Import Membership Records}
\usage{
btn_importRecords(Import, Current, conn)
}
\arguments{
\item{Import}{The data from the file being imported (Membership())}

\item{Current}{The Currently stored data (RV$Membership)}
}
\description{
Inserts or updates membership records from a CSV file into 
  the Membership table.
}

